fix: move http handler to composed skill
