+++ packages/packages.json
@@ -3,10 +3,10 @@
         "contract/valory/fpmm_deterministic_factory/0.1.0": "bafybeigjfuahxhb2y4q5ueayry55boaojcryoamwr6fshgayf5s762vpl4",
         "contract/valory/wxdai/0.1.0": "bafybeidalocwbhmbto6ii6adldtpcughtdt6j3v4tv36utevjk2wrdyqie",
         "contract/valory/fpmm/0.1.0": "bafybeiai2ruj27nnglvn7yc5atojyojo3fkmofw6wrjgz2ybps2uwdizx4",
-        "skill/valory/market_creation_manager_abci/0.1.0": "bafybeibl3buwan7qtzv3aqzuelpmmgydoy25gt4lbdnvxpettab3ltxp4m",
-        "skill/valory/market_maker_abci/0.1.0": "bafybeihbhymviqlx4ux2fhskua7zkpfupqqshy4tc5h2gsinva4idcqlcy",
-        "agent/valory/market_maker/0.1.0": "bafybeibcka3tam3a5e6mr4vwu55cwlolstvca5g54klnni4ijeqeka5jxa",
-        "service/valory/market_maker/0.1.0": "bafybeihowttwk2pzdpwg72tygbgebwyt2bscczcqeja4g3dpkdyp5sez6a"
+        "skill/valory/market_creation_manager_abci/0.1.0": "bafybeidgpctdcbxubjdlqzcrzr2qwtuocv236ngtjduabkmja7qvqr3fri",
+        "skill/valory/market_maker_abci/0.1.0": "bafybeietyfyi6qphajffab37bueo4seb2fdtptak6sneszk7ww7fjhwr5y",
+        "agent/valory/market_maker/0.1.0": "bafybeibbnx4azoeyuh2zsd6ebxvv2k2y3yvmuhfl4776p5pz4tzjb5uyuu",
+        "service/valory/market_maker/0.1.0": "bafybeiftf5j3vu7tab5uxfsidxrneayacfjbrgwejt5wk3qoyn4llgpkji"
     },
     "third_party": {
         "protocol/valory/contract_api/1.0.0": "bafybeidgu7o5llh26xp3u3ebq3yluull5lupiyeu6iooi2xyymdrgnzq5i",
